Revert UI styles for headers, fix styles and add responsive layouts for application dashboard banner
