+++ frontend/src/common/components/naviconsidebar/NavIconSideBar.tsx
@@ -1,9 +1,10 @@
-import {IDPS} from '../../types/idp'
-import './NavIconSideBar.scss'
-import { useAuth } from 'react-oidc-context'
-import OnRouteBCContext from "../../../common/authentication/OnRouteBCContext";
+import { useAuth } from "react-oidc-context";
 import React, { useContext } from "react";
-import { USER_AUTH_GROUP } from '../../../features/manageProfile/types/userManagement.d';
+
+import { IDPS } from "../../types/idp";
+import "./NavIconSideBar.scss";
+import OnRouteBCContext from "../../../common/authentication/OnRouteBCContext";
+import { USER_AUTH_GROUP } from "../../../features/manageProfile/types/userManagement.d";
 
 interface NavIconSideBarProps {
   children?: React.ReactNode;
@@ -13,17 +14,19 @@ interface NavIconSideBarProps {
  * Displays a sidebar with NavIcon buttons as children
  */
 export const NavIconSideBar = (props: NavIconSideBarProps) => {
-  const { children } = props
-  const { user } = useAuth()
-  const isIdir = user?.profile?.identity_provider === IDPS.IDIR
-  const { idirUserDetails } = useContext(OnRouteBCContext)
-  const isEofficer = idirUserDetails?.userAuthGroup === USER_AUTH_GROUP.EOFFICER
+  const { children } = props;
+  const { isAuthenticated, user } = useAuth();
+  const { idirUserDetails } = useContext(OnRouteBCContext);
+  const isIdir = user?.profile?.identity_provider === IDPS.IDIR;
+  const isEofficer = idirUserDetails?.userAuthGroup === USER_AUTH_GROUP.EOFFICER;
 
-  return (
+  const shouldShowSideBar = isAuthenticated && isIdir && !isEofficer;
+
+  return shouldShowSideBar ? (
     <div className="nav-icon-side-bar">
-      {isIdir && !isEofficer && children}
+      {children}
     </div>
-  )
+  ): null;
 }
