+++ frontend/src/features/permits/pages/TermOversize/components/dashboard/StartApplication.tsx
@@ -1,0 +1,54 @@
+import { useNavigate } from "react-router-dom";
+import { useState } from "react";
+import {
+  Box,
+  MenuItem,
+  SelectChangeEvent,
+  Button,
+} from "@mui/material";
+
+import "./StartApplication.scss";
+import { PERMIT_TYPE_CHOOSE_FROM_OPTIONS } from "../../../../constants/constants";
+import { SelectPermitType } from "./SelectPermitType";
+
+/**
+ *
+ * Code taken largely from MUI MenuList Composition
+ * https://mui.com/material-ui/react-menu/#menulist-composition
+ *
+ *
+ */
+export const StartApplication = () => {
+  const navigate = useNavigate();
+  const [chooseFrom, setChooseFrom] = useState("");
+  const handleChooseFrom = (event: SelectChangeEvent) => {
+    setChooseFrom(event.target.value as string);
+  };
+
+  const handleStartButtonClicked = () => {
+    navigate("/applications/permits");
+  };
+
+  return (
+    <Box className="start-application">
+      <SelectPermitType
+        value={chooseFrom}
+        label={"Select Permit Type"}
+        onChange={handleChooseFrom}
+        menuItems={PERMIT_TYPE_CHOOSE_FROM_OPTIONS.map((data) => (
+          <MenuItem key={data.value} value={data.value} selected={data.label === "Select"}>
+            {data.label}
+          </MenuItem>
+        ))}
+      />
+
+      <Button
+        className="start-application__btn"
+        variant="contained"
+        onClick={handleStartButtonClicked}
+      >
+        Start Application
+      </Button>
+    </Box>
+  );
+};
